New feature onClose added
Added onClose method to create trigger at instant when window is closed. This is useful in case when we need to execute some code  at the instant when user has closed window. We have added onClose as separate api so that it can be set if needed i.e. no change to current idle plugin method calls.

Co-Authored-By: KhurramKiani <68713560+KhurramKiani@users.noreply.github.com>